+++ database/migrations/2021_12_05_000003_create_subscriptions_table.php
@@ -18,13 +18,16 @@ class CreateSubscriptionsTable extends Migration
             $table->unsignedBigInteger('billable_id');
             $table->string('billable_type');
             $table->string('name');
-            $table->string('payfast_id')->unique();
-            $table->string('payfast_status');
-            $table->integer('payfast_plan');
-            $table->integer('quantity');
+            $table->integer('plan_id');
+            $table->string('token')->unique();
+            $table->string('status');
+            $table->string('payment_status')->nullable();
+            $table->string('subscription_status')->nullable();
+            $table->timestamp('next_bill_at')->nullable();
             $table->timestamp('trial_ends_at')->nullable();
             $table->timestamp('paused_from')->nullable();
             $table->timestamp('ends_at')->nullable();
+            $table->timestamp('cancelled_at')->nullable();
             $table->timestamps();
 
             $table->index(['billable_id', 'billable_type']);