+++ src/Billable.php
@@ -2,9 +2,8 @@
 
 namespace FintechSystems\Payfast;
 
-// use Laravel\Paddle\Concerns\ManagesCustomer;
-use Laravel\Paddle\Concerns\ManagesReceipts;
-use Laravel\Paddle\Concerns\PerformsCharges;
+use FintechSystems\Payfast\Concerns\ManagesReceipts;
+use FintechSystems\Payfast\Concerns\PerformsCharges;
 use FintechSystems\Payfast\Concerns\ManagesCustomer;
 use FintechSystems\Payfast\Concerns\ManagesSubscriptions;
 
@@ -12,8 +11,8 @@ trait Billable
 {
     use ManagesCustomer;
     use ManagesSubscriptions;
-    // use ManagesReceipts;
-    // use PerformsCharges;
+    use ManagesReceipts;
+    use PerformsCharges;
 
     /**
      * Get the default Paddle API options for the current Billable model.