+++ src/PayfastServiceProvider.php
@@ -31,19 +31,17 @@ class PayfastServiceProvider extends ServiceProvider
             __DIR__.'/../config/payfast.php', 'payfast'
         );
 
-        $this->app->bind('payfast', function () {
-            return new Payfast([
-                'merchant_id' => config('payfast.merchant_id'),
-                'merchant_key' => config('payfast.merchant_key'),
-                'passphrase' => config('payfast.passphrase'),
-                'testmode' => config('payfast.testmode'),
-                'return_url' => config('payfast.return_url'),
-                'cancel_url' => config('payfast.cancel_url'),
-                'notify_url' => config('payfast.notify_url'),
-                'card_update_link_css' => config('payfast.card_update_link_css'),
-                'card_updated_return_url' => config('payfast.card_updated_return_url'),
-            ]);
-        });
+        $this->app->bind('payfast', fn() => new Payfast([
+            'merchant_id' => config('payfast.merchant_id'),
+            'merchant_key' => config('payfast.merchant_key'),
+            'passphrase' => config('payfast.passphrase'),
+            'testmode' => config('payfast.testmode'),
+            'return_url' => config('payfast.return_url'),
+            'cancel_url' => config('payfast.cancel_url'),
+            'notify_url' => config('payfast.notify_url'),
+            'card_update_link_css' => config('payfast.card_update_link_css'),
+            'card_updated_return_url' => config('payfast.card_updated_return_url'),
+        ]));
         
     }
 }