+++ tests/Stubs/adhoc.payment.array.example.php
@@ -1,0 +1,27 @@
+<?php
+
+$array = [
+    "m_payment_id" => null,
+    "pf_payment_id" => "68774338",
+    "payment_status" => "COMPLETE",
+    "item_name" => "Order #1",
+    "item_description" => null,
+    "amount_gross" => "5.00",
+    "amount_fee" => "-2.50",
+    "amount_net" => "2.50",
+    "custom_str1" => null,
+    "custom_str2" => null,
+    "custom_str3" => null,
+    "custom_str4" => null,
+    "custom_str5" => null,
+    "custom_int1" => null,
+    "custom_int2" => null,
+    "custom_int3" => null,
+    "custom_int4" => null,
+    "custom_int5" => null,
+    "name_first" => null,
+    "name_last" => null,
+    "email_address" => null,
+    "merchant_id" => "13741656",
+    "signature" => "aba1635693fd5ac5c58385fa54f853f6"
+];