+++ tests/Stubs/fetch.subscription.array.example.php
@@ -1,0 +1,20 @@
+<?php
+
+$array = [
+    "code" => 200,
+    "status" => "success",
+    "data" => [
+        "response" => [
+            "amount" => 500,
+            "cycles" => 0,
+            "cycles_complete" => 2,
+            "frequency" => 3,
+            "run_date" => "2022-01-07T00:00:00+02:00",
+            "status" => 1,
+            "status_reason" => "",
+            "status_text" => "ACTIVE",
+            "token" => "21189d52-12eb-4108-9c0e-53343c7ac692"
+        ]
+    ]
+];
+