+++ tests/Stubs/subscription.canceled.array.example.php
@@ -1,27 +1,25 @@
 <?php
 
 $array = [
-
-    'm_payment_id' => 'new_ad_hoc_agreement_1',
-    'pf_payment_id' => '68613670',
+    'm_payment_id' => '2',
+    'pf_payment_id' => '68837816',
     'payment_status' => 'CANCELLED',
-    'item_name' => 'Setup Subscription Agreement',
-    'item_description' => 'Monthly Subscription',
+    'item_name' => 'Best Agent Test Monthly Subscription',
+    'item_description' => NULL,
     'custom_str1' => 'App\\Models\\User',
-    'custom_str2' => 'Monthly Subscription',
+    'custom_str2' => 'Monthly Plan',
     'custom_str3' => NULL,
     'custom_str4' => NULL,
     'custom_str5' => NULL,
     'custom_int1' => '1',
-    'custom_int2' => '1',
-    'custom_int3' => '1',
+    'custom_int2' => '2',
+    'custom_int3' => NULL,
     'custom_int4' => NULL,
     'custom_int5' => NULL,
-    'name_first' => 'Eugene van der Merwefn',
-    'name_last' => 'Eugene van der Merweln',
+    'name_first' => NULL,
+    'name_last' => NULL,
     'email_address' => 'eugene@vander.host',
     'merchant_id' => '13741656',
-    'token' => 'aab03980-76f7-4710-82b1-1d43cd739270',
-    'signature' => 'b7bded5bf89f06a280718d30a5bfb43b',
-
+    'token' => '50ad36e9-22a3-47fe-8385-5677e49ee025',
+    'signature' => '91d6b4f2edfacbd68b126719ea16cc34',
 ];